requirements.txt
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.12
ultralytics>=8.0.200
streamlit>=1.28.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
grad-cam
reportlab>=4.0.0
albumentations>=1.3.1
tqdm>=4.66.0
kaggle>=1.5.16
PyYAML>=6.0
seaborn>=0.12.0
fpdf2>=2.7.0


